# shared/generated/openapi-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Title",
    "description": "Title",
    "version": "3.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3333/api"
    }
  ],
  "paths": {
...
    "/auth/password/reset": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Reset password with a code and new password",
        "operationId": "resetPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "code": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  },
                  "confirmPassword": {
                    "type": "string"
                  }
                },
                "required": [
                  "code",
                  "newPassword",
                  "confirmPassword"
                ]
              }
            }
          },
          "description": "Reset password with a code and new password",
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Validation Error"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}